extract lexiconUI
